Rename types module to matrices
