+++ matrices/__init__.py
@@ -9,7 +9,7 @@ Version: 0.2.0
 Documentation: https://github.com/braedynl/matrices-py/wiki
 """
 
+from .matrices import *
 from .protocols import *
 from .rule import *
 from .shape import *
-from .types import *
